Added better feature for admin/user security

+++ app/Http/Controllers/UserController.php
@@ -27,9 +27,20 @@ public function showLists()
     return view('users.lists', compact('user', 'lists'));
 }
 
-    public function destroy()
+    public function destroy(Request $request)
     {
         $user = auth()->user();
+
+        if($user->isAdmin)
+        {
+            return back()->withErrors(['error' => 'Admin Konto kan ej raderas.']);
+        }
+
+        if (!$request->has('confirm') || $request->input('confirm') !== 'yes') {
+            return back()->withErrors(['error' => 'Bekräftelse krävs för att radera ditt konto.']);
+        }
+
+
         $user->delete();
         return redirect('/')->with('success', 'User deleted successfully');
     }
@@ -41,4 +52,5 @@ public function showLists()
         $request->session()->regenerateToken();
         return redirect('/');
     }
+
 }
